rename behavior -> image_level_behavior (contains I2n)
